cloud-functions/upload_fhir_bundle/requirements.txt
antlr4-python3-runtime==4.8
azure-core==1.25.0
azure-identity==1.10.0
azure-storage-blob==12.13.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cloudevents==1.4.0
coverage==6.4.3
cryptography==37.0.4
deprecation==2.1.0
fhirpathpy==0.1.0
Flask==2.1.3
functions-framework==3.1.0
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
hl7==0.4.5
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msal==1.18.0
msal-extensions==1.0.0
msrest==0.7.1
numpy==1.23.1
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
phdi @ git+https://github.com/CDCgov/phdi@37bcb41a2242694431f2fc3508f0ece0c4fb3256
phonenumbers==8.12.53
polling==0.3.2
portalocker==2.5.1
protobuf==4.21.4
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.3.5
pycparser==2.21
pydantic==1.9.2
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
smartystreets-python-sdk==4.11.6
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
watchdog==2.1.9
Werkzeug==2.2.0
zipp==3.8.1
